refactor(Project Structure): remove posts related stuff
