--- app/src/main/java/com/monstarlab/core/data/network/dtos/PostDto.kt
@@ -1,8 +1,0 @@
-package com.monstarlab.core.data.network.dtos
-
-data class PostDto(
-    val id: Int,
-    val userId: Int,
-    val title: String,
-    val body: String
-)
