--- app/src/main/java/com/monstarlab/core/domain/model/Post.kt
@@ -1,10 +1,0 @@
-package com.monstarlab.core.domain.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Post(
-    val id: Int,
-    val title: String,
-    val body: String,
-)
