+++ app/src/main/java/com/monstarlab/core/injection/modules/DataStoreModule.kt
@@ -1,11 +1,10 @@
-package com.monstarlab.injection.modules
+package com.monstarlab.core.injection.modules
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.SharedPreferencesMigration
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-import com.monstarlab.core.data.storage.PreferenceDataStore
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -15,7 +14,7 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-class DataModule {
+class DataStoreModule {
 
     private val Context.dataStore by preferencesDataStore(
         name = PreferenceDataStore.PREFS_NAME,
