+++ app/src/main/java/com/monstarlab/features/resources/data/ResourcesApi.kt
@@ -1,6 +1,5 @@
 package com.monstarlab.features.resources.data
 
-import com.monstarlab.core.data.network.responses.ResourcesResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
