+++ app/src/main/java/com/monstarlab/features/resources/data/ResourcesResponse.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.data.network.responses
+package com.monstarlab.features.resources.data
 
 import com.monstarlab.features.resources.data.ResourceDto
 import kotlinx.serialization.SerialName
