+++ app/src/main/java/com/monstarlab/features/resources/domain/ResourceRepository.kt
@@ -2,13 +2,13 @@ package com.monstarlab.features.resources.domain
 
 import com.monstarlab.arch.data.Repository
 import com.monstarlab.arch.extensions.mapSuccess
-import com.monstarlab.core.data.network.Api
 import com.monstarlab.core.data.network.dtos.toEntity
-import com.monstarlab.core.data.storage.ResourcePreferenceStore
+import com.monstarlab.features.resources.data.ResourcePreferenceStore
+import com.monstarlab.features.resources.data.ResourcesApi
 import javax.inject.Inject
 
 class ResourceRepository @Inject constructor(
-    private val api: Api,
+    private val api: ResourcesApi,
     private val resourcePreferenceStore: ResourcePreferenceStore
 ) : Repository() {
